scripts/requirements.txt
absl-py==0.8.0
alabaster==0.7.12
anykeystore==0.2
aspy.yaml==1.3.0
astor==0.8.0
attrs==19.1.0
Babel==2.7.0
backcall==0.1.0
bleach==3.1.0
blis==0.4.1
boto==2.49.0
boto3==1.9.207
botocore==1.12.207
certifi==2019.6.16
cffi==1.12.3
cfgv==2.0.1
chardet==3.0.4
Click==7.0
cycler==0.10.0
cymem==2.0.2
DAWG-Python==0.7.2
decorator==4.4.0
defusedxml==0.6.0
deprecation==2.0.7
docopt==0.6.2
docutils==0.14
entrypoints==0.3
fairseq==0.8.0
fastBPE==0.1.0
flake8==3.7.8
flake8-bugbear==19.8.0
flake8-docstrings==1.5.0
flake8-rst-docstrings==0.0.11
future==0.17.1
gast==0.3.1
gensim==3.8.0
gitdb2==2.0.6
GitPython==3.0.3
google-pasta==0.1.7
grpcio==1.23.0
h5py==2.10.0
hupper==1.8.1
hypothesis==3.88.3
identify==1.4.7
idna==2.8
imagesize==1.1.0
importlib-metadata==0.23
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.15.1
Jinja2==2.10.1
jmespath==0.9.4
joblib==0.13.2
jonze==0.0.20
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==5.3.1
jupyter-console==6.0.0
jupyter-core==4.5.0
Keras==2.3.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==2.6.11
MarkupSafe==1.1.1
matplotlib==3.1.1
mccabe==0.6.1
mistune==0.8.4
more-itertools==7.2.0
murmurhash==1.0.2
nbconvert==5.6.0
nbformat==4.4.0
nltk==3.4.5
nodeenv==1.3.3
notebook==6.0.1
num2words==0.5.10
numpy==1.17.0
oauthlib==3.1.0
onnx==1.5.0
opt-einsum==3.0.1
packaging==19.1
pandas==0.25.0
pandocfilters==1.4.2
parlai==0.1.0
parso==0.5.1
PasteDeploy==2.0.1
pbkdf2==1.3
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.0
pkg-resources==0.0.0
plac==0.9.6
plaster==1.0
plaster-pastedeploy==0.7
portalocker==1.5.1
pre-commit==1.18.3
preshed==3.0.2
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.9.1
ptyprocess==0.6.0
py-gfm==0.1.4
py-rouge==1.1
pyaml==19.4.1
pycodestyle==2.5.0
pycparser==2.19
pydocstyle==4.0.1
pyflakes==2.1.1
Pygments==2.4.2
pymorphy2==0.8
pymorphy2-dicts==2.4.393442.3710985
pyparsing==2.4.2
pyramid==1.10.4
pyramid-mailer==0.15.1
pyrsistent==0.15.4
pytext-nlp==0.2.2
python-crfsuite==0.9.6
python-dateutil==2.8.0
python3-openid==3.1.0
pytorch-pretrained-bert==0.6.2
pytz==2019.2
PyYAML==5.1.2
pyzmq==18.1.0
qtconsole==4.5.5
regex==2019.6.8
repoze.sendmail==4.4.1
requests==2.22.0
requests-mock==1.7.0
requests-oauthlib==1.2.0
restructuredtext-lint==1.3.0
s3transfer==0.2.1
sacrebleu==1.3.7
scikit-learn==0.21.3
scipy==1.3.1
Send2Trash==1.5.0
sh==1.12.14
six==1.12.0
sklearn-crfsuite==0.3.6
slize==0.0.21
smart-open==1.8.4
smmap2==2.0.5
snips-nlu==0.20.1
snips-nlu-parsers==0.4.1
snips-nlu-utils==0.9.1
snowballstemmer==2.0.0
spacy==2.2.1
Sphinx==2.2.0
sphinx-rtd-theme==0.4.3
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
SQLAlchemy==1.3.8
srsly==0.1.0
tabulate==0.8.3
tb-nightly==1.14.0a20190301
tensorboard==1.14.0
tensorboardX==1.8
tensorflow-estimator==1.14.0
tensorflow-gpu==2.0.0a0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
tf-estimator-nightly==1.14.0.dev2019080601
thinc==7.1.1
toml==0.10.0
torch==1.2.0
torchtext==0.4.0
tornado==6.0.3
tqdm==4.33.0
traitlets==4.3.2
transaction==2.4.0
translationstring==1.3
typing==3.7.4
typing-extensions==3.7.4
urllib3==1.25.3
velruse==1.1.1
venusian==1.2.0
virtualenv==16.7.5
wasabi==0.2.2
wcwidth==0.1.7
webencodings==0.5.1
WebOb==1.8.5
websocket-client==0.56.0
websocket-server==0.4
Werkzeug==0.15.6
widgetsnbextension==3.5.1
wrapt==1.11.2
WTForms==2.2.1
wtforms-recaptcha==0.3.2
zipp==0.6.0
zope.deprecation==4.4.0
zope.interface==4.6.0
zope.sqlalchemy==1.1
